requirements.txt
psycopg[binary,pool]
python-dotenv
pytest
pytest-cov